基本完成jvm模式的代码编写,下一步是日志的完善,node模式代码的编写以及监控中心的设计编码

+++ Kafka-Multithread/Kafka-Multithread-Protocol/src/main/java/org/kin/kafka/multithread/protocol/distributed/NodeMasterProtocol.java
@@ -8,6 +8,6 @@ import org.kin.kafka.multithread.domain.HealthReport;
  * 告诉Node关闭Container
  */
 public interface NodeMasterProtocol {
-    Boolean closeContainer(int containerId);
+    Boolean closeContainer(long containerId);
     void report(HealthReport report);
 }
